+++ custom_components/unifiprotect/binary_sensor.py
@@ -15,10 +15,7 @@ _LOGGER = logging.getLogger(__name__)
 
 DEPENDENCIES = ['unifiprotect']
 
-SCAN_INTERVAL = timedelta(seconds=5)
-
-ATTR_THUMBNAIL = "thumbnail"
-ATTR_LAST_MOTION = "last_motion"
+SCAN_INTERVAL = timedelta(seconds=3)
 
 # sensor_type [ description, unit, icon ]
 SENSOR_TYPES = {
@@ -54,8 +51,6 @@ class UfpBinarySensor(BinarySensorDevice):
         self._sensor_type = sensor_type
         self._nvrdata = nvrdata
         self._state = False
-        self._thumbnail = None
-        self._last_motion = None
         self._class = SENSOR_TYPES.get(self._sensor_type)[1]
         self._attr = SENSOR_TYPES.get(self._sensor_type)[2]
         self.remove_timer = None
@@ -78,8 +73,6 @@ class UfpBinarySensor(BinarySensorDevice):
 
         attrs[ATTR_ATTRIBUTION] = ATTRIBUTION
         attrs['brand'] = DEFAULT_BRAND
-        attrs[ATTR_THUMBNAIL] = self._thumbnail
-        attrs[ATTR_LAST_MOTION] = self._last_motion 
         attrs['friendly_name'] = self._name
 
         return attrs