+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Version 0.0.6
 * `Binary Motion`:
-  * Added new attribute Writes **last_motion** that shows time when the last motion has occured
+  * Changed default SCAN_INTERVAL to 3 seconds to optimize Motion Detection
 * `Camera`:
   * Added new service `camera.unifiprotect_save_thumbnail`. When calling this services the Thumbnail of the last recorded motion will be saved to disk, and could then be used in an automation, to send to a phone via the `notify` platform. Requires `entity_id` and `filename` as attributes.
 * `Core Module`